feat：新增大杂烩

+++ csagent-modules/csagent-application/src/main/java/com/csagent/application/domain/AppApplication.java
@@ -85,7 +85,7 @@ public class AppApplication extends TenantEntity {
     /**
      * 召回数量
      */
-    private Long topRank;
+    private Integer topRank;
 
     /**
      * 重排索引模型
@@ -95,17 +95,17 @@ public class AppApplication extends TenantEntity {
     /**
      * 记忆条数
      */
-    private Long memoryNum;
+    private Integer memoryNum;
 
     /**
      * 回复上限
      */
-    private Long maxReplyToken;
+    private Integer maxReplyToken;
 
     /**
      * 温度
      */
-    private Long temperature;
+    private Double temperature;
 
     /**
      * 是否显示引用
